Faster compilation time (especially for small programs)
The last step of the compilation consists in building the binary.
The problem is that last step rebuilds every time from source the file sk_standalone.
Because it includes headers from C++ libraries, compiling that file actually takes quite a bit of time.
So I changed the script to use the .o if it is there.

That bring the compilation time for a small program (like hello world) down from 10s to 5s on my machine.