Speedup of the  string literal lexer
Loading a file with large strings was taking a long time because
the implementation of substring was not very efficient (it was
using unsafe_get, which is O(n) on utf8 strings).

Also, the special character decoding made many calls to substring,
which is O(n) (again because of ut8).