First version of the skserver.
The big things that are missing right now are binaries so that
people can write skip code and iterate quickly on it.

It seems odd that this is missing so far into the project, but
the team was focused on a workflow that was originally going to
be integrated with HHVM. So all the focus was on that experience,
neglecting the command line experience.

Given that the HHVM support has been dropped. This diff is the
first version of what skip on the command line will look like.
It needs some love, but the core principles are there.

Before I get into the details, the most important thing to
understand is that skip is not meant to be used like a normal
command line compiler. We wrote skip in skip for a reason: we
wanted the compiler itself to be incremental!

In other words, when you modify a large project, rebuilding it
is very fast (much faster than the original build), because we
kept around all the intermediate values that did not change.

For that to work, you need skip to act like a server, that takes
commands like build/check/run etc ... So that the context from
one run to the next is not lost.

This is the first implementation of that server. Concretely what
does it implement:

- skserver
A bash script that establishes a connection (via named FIFOs) to
the binary skip_server. The binary takes commands from stdin and
spits out results in stdout. The skserver makes sure everything
is cleaned up etc ... The role of the skserver is to produce
llvm bitcode files (or error messages) for a given project.

- sk
A bash script that takes the commands and acts accordingly.

- prepare-release.sh
A script that takes all the binaries/libraries and other object
files that are necessary to build a skip program and puts them in
the directory dist/

- install.sh [dirname]
A script that installs skip, to the directory passed as a parameter
if one is provided, or to the default installation path if not.
Note that you might have to be root for the default path.

So what does the typical workflow look like?

Create your skip working directory:
$ mkdir myskip

Then write an sk file:
$ echo "fun main(): void { print_string(\"Hello World!\") }" > myskip/hello.sk

Then initialize the project (by creating a skip.project.json file):
$ sk init myskip

Then start the skserver:
$ sk start myskip

Then check that there are not errors in your program:
$ sk check myskip

Then build the project:
$ sk build myskip

The run it:
$ sk run myskip

Voila!

Some considerations:

There are a couple of races in the implementation. But I really think
those are normal. When you are writing a build system, I think everybody will
understand that things are going to go south if you modify the files at the
same time that you are trying to build things. So I didn't get out of my way
too hard to remove them.

I did not use inotify/fswatch etc ... On purpose. I really tried hard to keep
the dependencies of the bash script to pretty much zero. Granted, we now have
to stat all the files before every single build, and that's not going to scale
well. Eventually we will have to switch to an inotify based system. But when
is that going to be? "git" still stats everything ...

I did not write sk/skserver in SKIP, because there were too many system calls
that were not implemented. Well maybe that would have been the right time to
add them you may say? Not too sure about that. I really would like SKIP to be
compatible with windows long term. For that to be the case, we will need a
carefully thought out system API. I don't think that's something you write
quickly ... It will need some careful though in the design. So I think a wiser
plan is just to keep the system programming support minimum for now (open file,
close file etc ...), then add support for windows, then think hard about our
system programming API.

I considered writing sk/skserver in C/C++ too, I think it would have been fine
too.