Making the keys and the source of a reactive table generic. (#114)
Right now, the keys used by global tables are generic.
Reminder: reactive tables are tables where the keys are tracked by
the runtime the same way it would track and external dependency.

So the problem is that sometimes we will want to track keys that
are not strings. This diffs allows us to do that. What we do instead
of using a string is that we first intern it (to create a unique pointer
associated with the object), then take the interned pointer, turn it into
an int, and then turn that into a string.

Sure, we could probably have used the Int directly ... But this will
do, at least for now.