Make Skip use clang++-6.0 instead of just clang++
The problem is that sometimes there will be multiple versions of
clang installed on the system. When that is the case, there is
a way to update the symbolic links cleanly by using update_alternatives.

The problem with that is that it forces the user to commit to a version
of clang on that system, or forces constant back end forth betweeen versions.

So we are just going to commit to one version of clang, for now 6,
something more recent when we are ready.