removed Subprocess module
There was a primitive allowing to spawn a subprocess. Even though
that was useful, I am not sure I want to expose that API for the time
being, especially not in the standard library.

There are quite a few architectures out there, and I clearly won't
have time to maintain unix like APIs for them all.

I am ok with people making their own C++ library to do that, it's
just that I don't want to include that in the standard library.
At least not for now.