Simplified string hashing.
The hashing algorithm we were using had to match the one of HHVM,
because SKIP used to interop with Hack.

Given that we have given up on that interop, we don't need this
hashing function anymore. For example, PHP is case insensitive,
so the hashing algorithm had to reflect that.