First version of the Makefile
I have decided to move away from cmake/ninja to make.
The main reason is that I understand make better, but also because
it is available by default on more platforms than ninja.

The other thing that I am in the process of doing is getting rid of
all the C/C++ dependencies we had. I already got rid of: folly, boost,
libevent, libdouble-conversion ... to name a few.

So the good news is that we can now build skip with very few
dependencies:
  - libpcre for regular expressions
  - libicuuc/libicudata for unicode support
  - libunwind for backtraces and allocator profiling
  - libjemalloc for mallocx support

The bad news is that I am not done porting the test framework.
So we are going to live in that hybrid world where people who just
want to build skip can do so by using make, but people who want to
run the tests will have to install cmake/ninja among many other
dependencies.

It's a compromise I am ok with.