Disabling the logic zeroing gaps between fields in 32bits mode.
When initializing an object, the logic "zeros" the gaps between
the fields. I don't know why but I suspect it has to do with the
way interning works. The hash-consing of an object probably scans
the memory of the object at some point. The hash could give
different results if it included gaps with different values.

However, the interning in 32bits mode doesn't require this to
happen. So disabling it for now. It would probably be cleaner to
fix the math so that it also works in 32bits mode, but it's not
urgent.